+++ NEWS.md
@@ -1,3 +1,11 @@
+# oeli development
+
+* Added function `timed()` which evaluates an expression and interrupts the evaluation after a defined amount of seconds.
+
+* Added function `do.call_timed()` which measures the computation time of a `do.call()` call.
+
+* Added function `try_silent()` which tries to execute an expression and returns a string with the error message if the execution failed.
+
 # oeli 0.1.0
 
 * Initial CRAN submission.